Create AnimalToy2.0 (#16)
This version of the code shows the changes made to initialize the DF Player Mini with the sound file that is on there for the animal after the device is shaken. As well as the circuit now includes a transistor that will cut power to the speaker once it has had enough time to play the desired sound.